+++ src/lib.rs
@@ -16,39 +16,80 @@
 //!     ui.add(AutoCompleteTextEdit::new(
 //!        &mut self.text,
 //!        &self.inputs,
-//!        10,
 //!     ));
 //!   }
 //! }
 //! ````
-use egui::{Context, Id, Key, Modifiers, TextBuffer, TextEdit, Widget};
+use egui::{text::LayoutJob, Context, FontId, Id, Key, Modifiers, TextBuffer, TextEdit, Widget};
 use fuzzy_matcher::skim::SkimMatcherV2;
 use fuzzy_matcher::FuzzyMatcher;
 use std::cmp::{min, Reverse};
+
+/// Trait that can be used to modify the TextEdit
+type SetTextEditProperties = dyn FnOnce(TextEdit) -> TextEdit;
+
+/// An extension to the [`egui::TextEdit`] that allows for a dropdown box with autocomplete to popup while typing.  
+pub struct AutoCompleteTextEdit<'a> {
+    /// Contents of text edit passed into [`egui::TextEdit`]
+    text_field: &'a mut String,
+    /// Slice of strings to use as the search term
+    search: &'a [String],
+    /// A limit that can be placed on the maximum number of autocomplete suggestions shown
+    max_suggestions: usize,
+    /// If true, highlights the macthing indices in the dropdown
+    highlight: bool,
+    /// Used to set properties on the internal TextEdit
+    set_properties: Option<Box<SetTextEditProperties>>,
+}
 
 impl<'a> AutoCompleteTextEdit<'a> {
     /// Creates a new [`AutoCompleteTextEdit`].
     ///
     /// `text_field` - Contents of the text edit passed into [`egui::TextEdit`]
     /// `search` - Slice of strings to use as the search term
-    /// `max_suggestions` - Limit of max_suggestions to show in the drop down
-    pub fn new(text_field: &'a mut String, search: &'a [String], max_suggestions: usize) -> Self {
+    pub fn new(text_field: &'a mut String, search: &'a [String]) -> Self {
         Self {
             text_field,
             search,
-            max_suggestions,
+            max_suggestions: 10,
+            highlight: false,
+            set_properties: None,
         }
     }
 }
 
-/// An extension to the [`egui::TextEdit`] that allows for a dropdown box with autocomplete to popup while typing.  
-pub struct AutoCompleteTextEdit<'a> {
-    /// Contents of text edit passed into [`egui::TextEdit`]
-    text_field: &'a mut dyn TextBuffer,
-    /// Slice of strings to use as the search term
-    search: &'a [String],
-    /// A limit that can be placed on the maximum number of autocomplete suggestions shown
-    max_suggestions: usize,
+impl<'a> AutoCompleteTextEdit<'a> {
+    /// This determines the number of options appear in the dropdown menu
+    pub fn max_suggestions(mut self, max_suggestions: usize) -> Self {
+        self.max_suggestions = max_suggestions;
+        self
+    }
+    /// If set to true, characters will be highlighted in the dropdown to show the match
+    pub fn highlight_matches(mut self, highlight: bool) -> Self {
+        self.highlight = highlight;
+        self
+    }
+
+    /// Can be used to set the properties of the internal [`egui::TextEdit`]
+    /// # Example
+    /// ```rust
+    /// # use egui_autocomplete::AutoCompleteTextEdit;
+    /// # fn make_text_edit(mut search_field: String, inputs: Vec<String>) {
+    /// AutoCompleteTextEdit::new(&mut search_field, &inputs)
+    ///     .set_text_edit_properties(|text_edit: egui::TextEdit<'_>| {
+    ///         text_edit
+    ///             .hint_text("Hint Text")
+    ///             .text_color(egui::Color32::RED)
+    ///     });
+    /// # }
+    /// ```
+    pub fn set_text_edit_properties(
+        mut self,
+        set_properties: impl FnOnce(TextEdit) -> TextEdit + 'static,
+    ) -> Self {
+        self.set_properties = Some(Box::new(set_properties));
+        self
+    }
 }
 
 impl<'a> Widget for AutoCompleteTextEdit<'a> {
@@ -58,6 +99,8 @@ impl<'a> Widget for AutoCompleteTextEdit<'a> {
             text_field,
             search,
             max_suggestions,
+            highlight,
+            set_properties,
         } = self;
 
         let id = ui.next_auto_id();
@@ -70,7 +113,13 @@ impl<'a> Widget for AutoCompleteTextEdit<'a> {
             && ui.input_mut(|input| input.consume_key(Modifiers::default(), Key::ArrowUp));
         let down_pressed = state.focused
             && ui.input_mut(|input| input.consume_key(Modifiers::default(), Key::ArrowDown));
-        let text_response = TextEdit::singleline(text_field).ui(ui);
+
+        let mut text_edit = TextEdit::singleline(text_field);
+        if let Some(set_properties) = set_properties {
+            text_edit = set_properties(text_edit);
+        }
+
+        let text_response = text_edit.ui(ui);
         state.focused = text_response.has_focus();
 
         let matcher = SkimMatcherV2::default().ignore_case();
@@ -78,8 +127,8 @@ impl<'a> Widget for AutoCompleteTextEdit<'a> {
         let mut match_results = search
             .iter()
             .filter_map(|s| {
-                let score = matcher.fuzzy_match(s, text_field.as_str());
-                score.map(|x| (s, x))
+                let score = matcher.fuzzy_indices(s, text_field.as_str());
+                score.map(|(score, indices)| (s, score, indices))
             })
             .collect::<Vec<_>>();
         match_results.sort_by_key(|k| Reverse(k.1));
@@ -107,13 +156,27 @@ impl<'a> Widget for AutoCompleteTextEdit<'a> {
             text_field.replace(match_results[index].0)
         }
         egui::popup::popup_below_widget(ui, id, &text_response, |ui| {
-            for (i, (output, _)) in match_results.iter().take(max_suggestions).enumerate() {
-                let mut highlighed = if let Some(x) = state.selected_index {
+            for (i, (output, _, match_indices)) in
+                match_results.iter().take(max_suggestions).enumerate()
+            {
+                let mut selected = if let Some(x) = state.selected_index {
                     x == i
                 } else {
                     false
                 };
-                if ui.toggle_value(&mut highlighed, *output).clicked() {
+
+                let text = if highlight {
+                    highlight_matches(
+                        output,
+                        match_indices,
+                        ui.style().visuals.widgets.active.text_color(),
+                    )
+                } else {
+                    let mut job = LayoutJob::default();
+                    job.append(output, 0.0, egui::TextFormat::default());
+                    job
+                };
+                if ui.toggle_value(&mut selected, text).clicked() {
                     text_field.replace(output);
                 }
             }
@@ -133,6 +196,37 @@ impl<'a> Widget for AutoCompleteTextEdit<'a> {
 
         text_response
     }
+}
+
+/// Highlights all the match indices in the provided text
+fn highlight_matches(text: &&String, match_indices: &[usize], color: egui::Color32) -> LayoutJob {
+    let mut formatted = LayoutJob::default();
+    let mut it = (0..text.len()).peekable();
+    // Iterate through all indices in the string
+    while let Some(j) = it.next() {
+        let start = j;
+        let mut end = j;
+        let match_state = match_indices.contains(&start);
+        // Find all consecutive characters that have the same state
+        while let Some(k) = it.peek() {
+            if match_state == match_indices.contains(k) {
+                end += 1;
+                // Advance the iterator, we already peeked the value so it is fine to ignore
+                _ = it.next();
+            } else {
+                break;
+            }
+        }
+        // Format current slice based on the state
+        let format = if match_state {
+            egui::TextFormat::simple(FontId::default(), color)
+        } else {
+            egui::TextFormat::default()
+        };
+        let slice = &text[start..=end];
+        formatted.append(slice, 0.0, format);
+    }
+    formatted
 }
 
 /// Stores the currently selected index in egui state
@@ -214,7 +308,10 @@ mod test {
     }
     #[test]
     fn decrement_index() {
-        let mut state = AutoCompleteTextEditState::default();
+        let mut state = AutoCompleteTextEditState {
+            selected_index: Some(1),
+            ..Default::default()
+        };
         state.selected_index = Some(1);
         state.update_index(false, false, 10, 10);
         assert_eq!(Some(1), state.selected_index);
@@ -222,5 +319,27 @@ mod test {
         assert_eq!(Some(0), state.selected_index);
         state.update_index(false, true, 10, 10);
         assert_eq!(None, state.selected_index);
+    }
+    #[test]
+    fn highlight() {
+        let text = String::from("Test123");
+        let match_indices = vec![1, 5, 6];
+        let layout = highlight_matches(&&text, &match_indices, egui::Color32::RED);
+        assert_eq!(4, layout.sections.len());
+        let sec1 = layout.sections.get(0).unwrap();
+        assert_eq!(sec1.byte_range, 0..1);
+        assert_ne!(sec1.format.color, egui::Color32::RED);
+
+        let sec2 = layout.sections.get(1).unwrap();
+        assert_eq!(sec2.byte_range, 1..2);
+        assert_eq!(sec2.format.color, egui::Color32::RED);
+
+        let sec3 = layout.sections.get(2).unwrap();
+        assert_eq!(sec3.byte_range, 2..5);
+        assert_ne!(sec3.format.color, egui::Color32::RED);
+
+        let sec4 = layout.sections.get(3).unwrap();
+        assert_eq!(sec4.byte_range, 5..7);
+        assert_eq!(sec4.format.color, egui::Color32::RED);
     }
 }