+++ CHANGELOG.md
@@ -7,6 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2023-08-16 
+
+### Breaking
+- API has moved to a builder pattern, with only the required parameters in the `new` function
+    - Max suggestions is no longer in the `new` function. To get the same behavior change usage as follows  
+        **Old**
+        ```rust
+        AutoCompleteTextEdit::new(&mut text, &inputs, max_suggestions);
+        ```
+        **New**
+        ```rust
+        AutoCompleteTextEdit::new(&mut text, &inputs).max_suggestions(max_suggestions);
+        ```
+
+### Added
+- API to change max_suggestions `.max_suggestions(usize)`
+- API to enable highlighting of the matched characters `.highlight_matches(bool)`
+- API to modify the internal egui TextEdit `.set_text_edit_properties`
+    - See docs for example for this
+
+[0.2.0]: https://github.com/JakeHandsome/egui_autocomplete/compare/0.1.2...0.2.0
+
 ## [0.1.2] - 2023-08-07
 
 ### Added